closes #68; move headers to 'nanotime' dir; put 'period' and 'interval' in namespace; make 'pseudovector' more flexible
